Split authentication module in sessions and headers
