+++ lib/gamora.rb
@@ -10,5 +10,10 @@ module Gamora
 
   autoload :User, "gamora/user"
   autoload :Client, "gamora/client"
-  autoload :Authentication, "gamora/authentication"
+
+  module Authentication
+    autoload :Base, "gamora/authentication/base"
+    autoload :Headers, "gamora/authentication/headers"
+    autoload :Session, "gamora/authentication/session"
+  end
 end
