+++ lib/gamora/configuration.rb
@@ -13,7 +13,6 @@ module Gamora
     mattr_accessor :default_scope, default: "openid profile email"
     mattr_accessor :default_prompt, default: nil
     mattr_accessor :default_strategy, default: "default"
-    mattr_accessor :access_token_source, default: "session"
 
     def setup
       yield(self) if block_given?
